+++ src/components/TextInput/TextInput.tsx
@@ -11,26 +11,52 @@ import React, {
 } from 'react';
 import styled, { css } from 'styled-components';
 import Icon from '@mdi/react';
-import { mdiClose } from '@mdi/js';
 import debounce from 'lodash.debounce';
+import { mdiClose } from '@mdi/js';
+import { darken } from 'polished';
 import { Div, Input as InputElement, TextArea } from '../../htmlElements';
 import { SubcomponentPropsType, StyledSubcomponentType } from '../commonTypes';
 import { useAnalytics, useTheme } from '../../context';
 import { disabledStyles } from '../../utils/color';
+import variants from '../../enums/variants';
+
+export type TextInputContainerProps = {
+  disabled?: boolean;
+  isValid?: boolean;
+  variant?: variants;
+};
 
 const Container = styled(Div)`
-  ${({ disabled = false, isValid }: { disabled?: boolean; isValid?: boolean }) => {
+  ${({ disabled = false, isValid, variant = variants.outline }: TextInputContainerProps) => {
     const { colors } = useTheme();
+    const borderColor = isValid === false ? colors.destructive : colors.grayMedium;
     return `
-      border 1px solid ${isValid === false ? colors.destructive : colors.grayMedium};
       min-width: 10rem;
       position: relative;
       display: flex;
       flex-flow: row;
       border-radius: 0.25em;
-      background-color: ${colors.background};
+      border: ${
+        variant === variants.outline ? `1px solid ${borderColor}` : '1px solid transparent'
+      };
+
+      &:focus-within {
+        outline: none;
+        box-shadow: 0 0 5px 0.150rem ${colors.tertiary};
+      }
+      
+      ${
+        variant === variants.fill
+          ? `border-bottom: 1px solid ${borderColor}; 
+            border-bottom-left-radius: 0; 
+            border-bottom-right-radius: 0;`
+          : ''
+      }
+      background-color: ${
+        variant === variants.fill ? darken(0.1, colors.background) : colors.background
+      };
       ${disabled ? disabledStyles() : ''}
-  `;
+    `;
   }}
 `;
 
@@ -44,9 +70,6 @@ const TextInputContainer = styled(InputElement)`
       font-size: 1em;
       padding: 0.5rem;
       background-color: ${colors.transparent};
-      &:focus {
-        outline: none;
-        box-shadow: 0 0 5px 0.150rem ${colors.tertiary};
   `;
   }}
 `;
@@ -63,9 +86,6 @@ const TextAreaInputContainer = styled(TextArea)`
       padding: .5rem;
       background-color: ${colors.transparent};
       resize: ${multiLineIsResizable ? 'both' : 'none'};
-      &:focus {
-        outline: none;
-        box-shadow: 0 0 5px 0.150rem ${colors.tertiary};
     `;
   }}
 `;
@@ -76,6 +96,7 @@ const IconContainer = styled(Div)`
     return `
       padding: 0.5em;
       height: 100%;
+      display: flex;
       align-items: center;
       justify-content: center;
       color: ${colors.grayMedium};
@@ -117,6 +138,7 @@ export type TextInputProps = InputHTMLAttributes<HTMLInputElement> &
     isValid?: boolean;
     isMultiline?: boolean;
     errorMessage?: string;
+    variant?: variants;
     debounceInterval?: number;
     multiLineIsResizable?: boolean;
     maxLength?: number;
@@ -142,11 +164,15 @@ export type TextInputProps = InputHTMLAttributes<HTMLInputElement> &
     characterCountRef?: React.RefObject<HTMLDivElement>;
   };
 
-const createIcon = (StyledIconContainer: StyledSubcomponentType, iconPrefix: ReactNode) => {
+const createIcon = (
+  StyledIconContainer: StyledSubcomponentType,
+  iconPrefix: ReactNode,
+  iconProps: SubcomponentPropsType,
+) => {
   if (typeof iconPrefix === 'string') {
     return (
-      <StyledIconContainer>
-        <Icon aria-hidden="true" size="1rem" path={iconPrefix} />
+      <StyledIconContainer {...iconProps}>
+        <Icon aria-hidden="true" size="1em" path={iconPrefix} />
       </StyledIconContainer>
     );
   }
@@ -164,6 +190,7 @@ const TextInput = ({
   isValid = true,
   isMultiline,
   errorMessage,
+  variant = variants.outline,
   debounceInterval = 8,
   multiLineIsResizable,
   maxLength,
@@ -235,10 +262,11 @@ const TextInput = ({
     <StyledContainer
       disabled={nativeHTMLAttributes.disabled}
       isValid={isValid}
+      variant={variant}
       ref={containerRef}
       {...containerProps}
     >
-      {iconPrefix && createIcon(StyledIconContainer, iconPrefix)}
+      {iconPrefix && createIcon(StyledIconContainer, iconPrefix, iconContainerProps)}
       <InputComponent
         // Set default values above nativeHTMLAttributes
         type="text"